Changed backup/restore to use encrypted zip files. It now also gets the filename and password from the arguments.

+++ backup.sh
@@ -1,13 +1,16 @@
 #! /bin/bash
-time=$(date "+%Y-%m-%d_%H-%M-%S")
-dirname=/tmp/hydrominder_backup_$time
-mkdir -p $dirname
-INSTALL_DIR="/var/lib/hydrominder"
-cd $dirname
-sudo docker exec -i postgres /usr/local/bin/pg_dumpall -U hydrominder > postgres-backup.sql
-sudo cp -r $INSTALL_DIR/ssl .
-sudo cp -r $INSTALL_DIR/var .
-BACKUP_DIR="$INSTALL_DIR/backups"
-mkdir -p $BACKUP_DIR
-tar czf $BACKUP_DIR/$time.tar.gz *
-rm -r $dirname
+if [ "$#" -ne 1 ]; then
+    echo "Illegal number of parameters. Usage: ./backup.sh filename.zip password"
+else
+    dirname=/tmp/hydrominder_backup_$1
+    mkdir -p $dirname
+    INSTALL_DIR="/var/lib/hydrominder"
+    cd $dirname
+    sudo docker exec -i postgres /usr/local/bin/pg_dumpall -U hydrominder > postgres-backup.sql
+    sudo cp -r $INSTALL_DIR/ssl .
+    sudo cp -r $INSTALL_DIR/var .
+    BACKUP_DIR="$INSTALL_DIR/backups"
+    mkdir -p $BACKUP_DIR
+    zip -e -P $2 -r $BACKUP_DIR/$1 *
+    rm -r $dirname
+fi
